group id - route fixed and wrapped Activity tab in PageWrapperView
