+++ app/(home)/_layout.tsx
@@ -3,6 +3,7 @@ import { Stack } from 'expo-router';
 export default function Layout() {
   return (
     <Stack screenOptions={{ headerShown: false }}>
+      <Stack.Screen name='(tabs)' />
       <Stack.Screen name='(setting)' />
       <Stack.Screen name='(modal)' />
     </Stack>
